Separate client functions in different files
