+++ client.go
@@ -1,11 +1,9 @@
 package nanogit
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
 	"strings"
@@ -126,130 +124,6 @@ func (c *httpClient) getLogger(ctx context.Context) Logger {
 	}
 
 	return c.logger
-}
-
-// uploadPack sends a POST request to the git-upload-pack endpoint.
-// This endpoint is used to fetch objects and refs from the remote repository.
-func (c *httpClient) uploadPack(ctx context.Context, data []byte) ([]byte, error) {
-	body := bytes.NewReader(data)
-
-	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/git-upload-pack.
-	// See: https://git-scm.com/docs/protocol-v2#_http_transport
-	u := c.base.JoinPath("git-upload-pack").String()
-
-	logger := c.getLogger(ctx)
-
-	logger.Info("UploadPack", "url", u, "requestBody", string(data))
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u, body)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Content-Type", "application/x-git-upload-pack-request")
-	c.addDefaultHeaders(req)
-
-	res, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer res.Body.Close()
-
-	if res.StatusCode < 200 || res.StatusCode >= 300 {
-		return nil, fmt.Errorf("got status code %d: %s", res.StatusCode, res.Status)
-	}
-
-	responseBody, err := io.ReadAll(res.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	logger.Info("UploadPack", "status", res.StatusCode, "statusText", res.Status, "responseBody", string(responseBody), "requestBody", string(data), "url", u)
-
-	return responseBody, nil
-}
-
-// receivePack sends a POST request to the git-receive-pack endpoint.
-// This endpoint is used to send objects to the remote repository.
-func (c *httpClient) receivePack(ctx context.Context, data []byte) ([]byte, error) {
-	body := bytes.NewReader(data)
-
-	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/git-receive-pack.
-	// See: https://git-scm.com/docs/protocol-v2#_http_transport
-	u := c.base.JoinPath("git-receive-pack")
-	logger := c.getLogger(ctx)
-	logger.Info("GitReceivePack", "url", u.String())
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u.String(), body)
-	if err != nil {
-		return nil, err
-	}
-
-	c.addDefaultHeaders(req)
-	req.Header.Add("Content-Type", "application/x-git-receive-pack-request")
-	req.Header.Add("Accept", "application/x-git-receive-pack-result")
-
-	res, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer res.Body.Close()
-
-	if res.StatusCode < 200 || res.StatusCode >= 300 {
-		return nil, fmt.Errorf("got status code %d: %s", res.StatusCode, res.Status)
-	}
-
-	responseBody, err := io.ReadAll(res.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	logger.Info("ReceivePack", "status", res.StatusCode, "statusText", res.Status, "responseBody", string(responseBody), "requestBody", string(data), "url", u.String())
-
-	return responseBody, nil
-}
-
-// smartInfo sends a GET request to the info/refs endpoint.
-func (c *httpClient) smartInfo(ctx context.Context, service string) ([]byte, error) {
-	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/info/refs.
-	// The ?service=git-upload-pack is documented in the protocol-v2 spec. It also implies elsewhere that ?svc is also valid.
-	// See: https://git-scm.com/docs/http-protocol#_smart_clients
-	// See: https://git-scm.com/docs/protocol-v2#_http_transport
-	u := c.base.JoinPath("info/refs")
-
-	query := make(url.Values)
-	query.Set("service", service)
-	u.RawQuery = query.Encode()
-
-	logger := c.getLogger(ctx)
-	logger.Info("SmartInfo", "url", u.String())
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	c.addDefaultHeaders(req)
-
-	res, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer res.Body.Close()
-	if res.StatusCode < 200 || res.StatusCode >= 300 {
-		return nil, fmt.Errorf("got status code %d: %s", res.StatusCode, res.Status)
-	}
-
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	logger.Info("SmartInfo", "status", res.StatusCode, "statusText", res.Status, "body", string(body))
-
-	return body, nil
 }
 
 // NewHTTPClient creates a new Git client for the specified repository URL.
