+++ receivepack.go
@@ -1,0 +1,50 @@
+package nanogit
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io"
+	"net/http"
+)
+
+// receivePack sends a POST request to the git-receive-pack endpoint.
+// This endpoint is used to send objects to the remote repository.
+func (c *httpClient) receivePack(ctx context.Context, data []byte) ([]byte, error) {
+	body := bytes.NewReader(data)
+
+	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/git-receive-pack.
+	// See: https://git-scm.com/docs/protocol-v2#_http_transport
+	u := c.base.JoinPath("git-receive-pack")
+	logger := c.getLogger(ctx)
+	logger.Info("GitReceivePack", "url", u.String())
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u.String(), body)
+	if err != nil {
+		return nil, err
+	}
+
+	c.addDefaultHeaders(req)
+	req.Header.Add("Content-Type", "application/x-git-receive-pack-request")
+	req.Header.Add("Accept", "application/x-git-receive-pack-result")
+
+	res, err := c.client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	defer res.Body.Close()
+
+	if res.StatusCode < 200 || res.StatusCode >= 300 {
+		return nil, fmt.Errorf("got status code %d: %s", res.StatusCode, res.Status)
+	}
+
+	responseBody, err := io.ReadAll(res.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	logger.Info("ReceivePack", "status", res.StatusCode, "statusText", res.Status, "responseBody", string(responseBody), "requestBody", string(data), "url", u.String())
+
+	return responseBody, nil
+}
