+++ smartinfo.go
@@ -1,0 +1,51 @@
+package nanogit
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"net/http"
+	"net/url"
+)
+
+// smartInfo sends a GET request to the info/refs endpoint.
+func (c *httpClient) smartInfo(ctx context.Context, service string) ([]byte, error) {
+	// NOTE: This path is defined in the protocol-v2 spec as required under $GIT_URL/info/refs.
+	// The ?service=git-upload-pack is documented in the protocol-v2 spec. It also implies elsewhere that ?svc is also valid.
+	// See: https://git-scm.com/docs/http-protocol#_smart_clients
+	// See: https://git-scm.com/docs/protocol-v2#_http_transport
+	u := c.base.JoinPath("info/refs")
+
+	query := make(url.Values)
+	query.Set("service", service)
+	u.RawQuery = query.Encode()
+
+	logger := c.getLogger(ctx)
+	logger.Info("SmartInfo", "url", u.String())
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, u.String(), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	c.addDefaultHeaders(req)
+
+	res, err := c.client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	defer res.Body.Close()
+	if res.StatusCode < 200 || res.StatusCode >= 300 {
+		return nil, fmt.Errorf("got status code %d: %s", res.StatusCode, res.Status)
+	}
+
+	body, err := io.ReadAll(res.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	logger.Info("SmartInfo", "status", res.StatusCode, "statusText", res.Status, "body", string(body))
+
+	return body, nil
+}
